+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+[4.0.0] - 2018-04-02
+--------------------
+- [x] Remove `respond` callback from `listen`
+- [x] Move to `trid` for ids
+- [x] Formattable logger
+
+
 [3.3.0] - 2018-03-30
 --------------------
 - [x] Add custom error class #8